Rest controller and bug fix

+++ src/main/java/com/did/todo/model/ItemRepository.java
@@ -2,6 +2,6 @@ package com.did.todo.model;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ItemRepository extends PagingAndSortingRepository<Item, Long> {
+public interface ItemRepository extends PagingAndSortingRepository<Item, Integer> {
 
 }
